feat: Report type 생성
