+++ src/app/mypage/report-list/_components/ReportListItem.tsx
@@ -1,12 +1,5 @@
 import { Button } from '@/components/ui/button';
-
-export interface Report {
-  reportId: number;
-  reporterId: number;
-  reportedId: number;
-  reportedNickname: string;
-  content: string;
-}
+import { Report } from '@/types/report';
 
 interface ReportListItemProps {
   data: Report;
