+++ src/types/report.ts
@@ -1,0 +1,7 @@
+export interface Report {
+  reportId: number;
+  reporterId: number;
+  reportedId: number;
+  reportedNickname: string;
+  content: string;
+}
